requirements.txt
kafka-python==2.0.2
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
dash==2.14.1
plotly==5.17.0
dash-bootstrap-components==1.5.0
